Add initial support for 'ephemeral' memory region
The idea behind this region is that temporary values constructed by a
query can be removed all at once after a query has finished.

The current version treats all terms constructed by unification as
ephemeral, but it doesn't yet treat the query terms as ephemeral.
Treating query terms as ephemeral requires a refactor to the parser so
that hashconsing is delayed until after the initial parse.